Relocating the DeckPickerWidget-specific files to a dedicated package
