+++ AnkiDroid/src/main/AndroidManifest.xml
@@ -523,7 +523,7 @@
         <!-- A widget that displays a few decks's name and number of cards to review on the Android home screen.
              The way to add it depends on the phone. It usually consists in a long press on the screen, followed by finding a "widget" button"-->
         <receiver
-            android:name="com.ichi2.widget.DeckPickerWidget"
+            android:name="com.ichi2.widget.deckpicker.DeckPickerWidget"
             android:label="@string/deck_picker_widget_description"
             android:exported="false"
             >
@@ -540,7 +540,7 @@
              It is opened when adding a new deck picker widget and
              by configuration button which appears when the widget is hold or resized.-->
         <activity
-            android:name="com.ichi2.widget.DeckPickerWidgetConfig"
+            android:name="com.ichi2.widget.deckpicker.DeckPickerWidgetConfig"
             android:exported="false">
             <intent-filter>
                 <action android:name="android.appwidget.action.APPWIDGET_CONFIGURE" />
